+++ src/core/utils/chains.ts
@@ -1,25 +1,9 @@
 import { AddressZero } from '@ethersproject/constants';
-import {
-  arbitrum,
-  arbitrumGoerli,
-  base,
-  baseGoerli,
-  bsc,
-  bscTestnet,
-  goerli,
-  mainnet,
-  optimism,
-  optimismGoerli,
-  polygon,
-  polygonMumbai,
-  zora,
-  zoraTestnet,
-} from '@wagmi/chains';
 import { getNetwork } from '@wagmi/core';
-import { type Chain, sepolia } from 'wagmi';
+import { mainnet } from 'wagmi';
 
-import { NATIVE_ASSETS_PER_CHAIN } from '~/core/references';
-import { ChainId, ChainName, ChainNameDisplay } from '~/core/types/chains';
+import { NATIVE_ASSETS_PER_CHAIN, SUPPORTED_CHAINS } from '~/core/references';
+import { ChainId, ChainName } from '~/core/types/chains';
 
 import { customRPCsStore } from '../state/customRPC';
 import { AddressOrEth } from '../types/assets';
@@ -27,34 +11,6 @@ import { AddressOrEth } from '../types/assets';
 import { getDappHost } from './connectedApps';
 import { isCustomNetwork } from './customNetworks';
 import { isLowerCaseMatch } from './strings';
-
-export const SUPPORTED_MAINNET_CHAINS: Chain[] = [
-  mainnet,
-  polygon,
-  optimism,
-  arbitrum,
-  base,
-  zora,
-  bsc,
-].map((chain) => ({ ...chain, name: ChainNameDisplay[chain.id] }));
-
-export const SUPPORTED_CHAINS: Chain[] = [
-  mainnet,
-  polygon,
-  optimism,
-  arbitrum,
-  base,
-  zora,
-  bsc,
-  goerli,
-  sepolia,
-  optimismGoerli,
-  bscTestnet,
-  polygonMumbai,
-  arbitrumGoerli,
-  baseGoerli,
-  zoraTestnet,
-].map((chain) => ({ ...chain, name: ChainNameDisplay[chain.id] }));
 
 export const getSupportedChainsWithHardhat = () => {
   const { chains } = getNetwork();
